pyproject.toml
[tool.poetry]
name = "pushbib"
version = "0.1.0"
description = "A web application for saving reference information to a database and exporting the information from said website in a BibTex format."
authors = ["PushIT"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.2"
gunicorn = "^20.1.0"
flask-sqlalchemy = "^3.0.2"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
pylint = "^2.15.6"
requests = "^2.28.1"
invoke = "^1.7.3"
bibtexparser = "^1.4.0"
flask-babel = "^2.0.0"
babel = "^2.11.0"

[tool.poetry.dev-dependencies]
robotframework = "^6.0.1"
pytest = "^7.2.0"
autopep8 = "^2.0.0"
robotframework-seleniumlibrary = "^6.0.0"
coverage = "^6.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
